PropAnalyzer/ai/requirements.txt
# FastAPI و سرور
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# تحلیل داده و ML
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
joblib==1.3.2
scipy==1.11.4

# دیتابیس
psycopg2-binary==2.9.9
python-decouple==3.8

# ابزارهای کمکی
python-multipart==0.0.6